+++ src/config/client.rs
@@ -2,10 +2,12 @@ use super::{CommonConfig, NetworkConfig};
 use serde::{Deserialize, Serialize};
 
 mod action;
+mod connect;
 mod launch;
 mod repl;
 
 pub use action::*;
+pub use connect::*;
 pub use launch::*;
 pub use repl::*;
 
@@ -16,6 +18,7 @@ pub struct ClientConfig {
     pub common: CommonConfig,
 
     pub action: ClientActionConfig,
+    pub connect: ClientConnectConfig,
     pub launch: ClientLaunchConfig,
     pub repl: ClientReplConfig,